[device-keygen-module] Move device keygen into device/keygen.rs
Give the device's keygen domain the same module home the coordinator gets in PR #458 and restoration already has on the device side: keygen::State for the tmp maps (composed as FrostSigner.keygen), the KeyGenPhase structs, the KeyGen dispatch with per-variant #[inline(never)] handlers (preserving the stack-frame win, keygen path 5552 bytes peak), and keygen_ack. Keygen only. Zero logic change — moved code proven token-identical to the pre-move source modulo five enumerated mechanical transforms.
